Finished Blind Date (I think)
-Added revealButton function to MessageMgr that is called when the
reveal profiles button is pressed by a user in a blind date. It checks
if the other user already pressed the button, and updates the DB
accordingly.

- Added function hasUserPressedReveal which returns true or false based
on whether or not the logged in user has pressed the reveal button on a
blind date.

- Added ConversationLoader function to MessageMgr that loads everything
on the conversations page except the heading and back button.

- Added blindDateEligible function to MessageMgr that checks if a user
is eligible to sign up for a blind date. If they're not, a suitable
error message is returned.

- Added getOtherUser function to MessageMgr that is called when a user
presses the Send Message button on another user's profile. It finds
whether or not a conversation between the two users already exists or
not and redirects accordingly.

- Tidied up conversationPage and blindDatePage

- Moved form for send message button from profilePage to
sendMessageButton function in MessageMgr